Introduce PlaceConfig and AnalyticsConfig instead of Simple config
