+++ src/Config/PlacesConfig.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Algolia\AlgoliaSearch\Config;
+
+class PlacesConfig extends AbstractConfig
+{
+    public static function create($appId, $apiKey)
+    {
+        $config = array(
+            'appId' => $appId,
+            'apiKey' => $apiKey,
+        );
+
+        return new static($config);
+    }
+}
